create page folder, relocate page-components
