+++ src/components/App/App.jsx
@@ -5,10 +5,10 @@ import ProtectedRoute from '../ProtectedRoute/ProtectedRoute.jsx';
 import NotFound from '../NotFound/NotFound.jsx';
 import Boards from '../Boards/Boards.jsx';
 import { endpoint } from '../../constants/constantsEndpointRoute.js';
-import Auth from '../Auth/Auth.jsx';
-import Register from '../Register/Register.jsx';
+import Auth from '../../pages/Auth/Auth.jsx';
+import Register from '../../pages/Register/Register.jsx';
 import { boardsList } from '../../constants/boardsList.js';
-import PersonalArea from '../PersonalArea/PersonalArea.jsx';
+import PersonalArea from '../../pages/PersonalArea/PersonalArea.jsx';
 import AdminPanel from '../AdminPanel/AdminPanel.jsx';
 
 function App() {
