+++ src/components/MainPage/MainPage.jsx
@@ -1,7 +1,7 @@
 import './MainPage.scss';
-import PersonalArea from '../PersonalArea/PersonalArea.jsx';
+import PersonalArea from '../../pages/PersonalArea/PersonalArea.jsx';
 import AdminPanel from '../AdminPanel/AdminPanel.jsx';
-import Register from '../Register/Register.jsx';
+import Register from '../../pages/Register/Register.jsx';
 
 function MainPage() {
   const status = !true;
